Fixes the crash part of Bug 1305. I am logging a warning message that should be shown to the user somehow whenever this occurs, but I am not sure of the best mechanism for doing that since the change is down in the OSDocument constructor.
The osp in the osm file that crashes is stamped as Version 0.11.5, but contains database elements of 0.11.6. That mismatch was the specific cause of the crash. I do not know how this out-of-syncness happened, but I suspect that a mid-iteration build may have been involved ... ?

Thus, I have fixed the crash part of this bug, but in the course of doing that I am dropping some user data, and there may be an underlying issue lurking about.

[delivers #52494241]