Organize HVAC views and controllers
Shuffled the chairs to better support the growing reach of the HVAC tab,
including controls, refrigeration, hvac layout, and hot water systems.
This change avoids keeping all of that in memory at one time, instead
allocating views and controllers for only the type of HVAC system you
are working on.  Lower memory, faster refresh, and easier to read code.