Get OSGridController templates compiling
Needed to templatize the return type, since you are trying to return
something specific (ThermalZone) into something generic (ModelObject).

All of the previous examples that I worked on with Kyle were always
returning something specific (int, bool, etc).

Not a big deal or difficult, and I was able to clean up the call to the
addColumn function by dropping the template parameters as well.

@kbenne should probably review to make sure that it's still doing what
he intended it to do, but I believe it is.

@evanweaver.