[#65315174] Place ruby libs in shared location in linux
Ruby libs now install into a (somewhat) self contained location at
<prefix>/lib/openstudio-<openstudioversion>/ruby/<rubyversion>/

Symlinks are installed which point from
<prefix>/lib/ruby/site_ruby/<rubyversion>/ to the installed location.
This allows for easy placement of additional ruby installations if
needed and makes the ruby .so finding easier to accomplish.

I also updated the ruby PATH to be smart about symlinks and point to the
actual location instead of the symlink location.

I have one further test to make, but I think this is ready. @axelstudios
might need to review it because of the linux folder layout changes. I
tried to keep other platforms the same that they were because they
already make self contained installs of ruby.

@nllong