Add cascade system to library
Add a cascade system to the library, meaning this system is named
accordingly and it has a cascade condenser already attached to the
supply side.  Teach RefrigerationController how to deal with this
situation.  Prior to this commit only empty RefrigerationSystems were
handled well, and in that case when an empty system was dropped on
another system as a load, a cascade condenser was made automatically to
join the two.  This is still true, just adding support for systems that
already have cascade condenser from the library.