Fix merge related errors introduced by 73802200-SDDEnhancements
These are corrections to merge errors related to
73802200-SDDEnhancements being merged to develop.  The main difficulty
is getting that work up to speed with the dependency update that is now
included in the develop branch.  Normally this would be worked out on
the SDDEnhancements branch, however in this case we do not want the
changes of develop (new dependencies) introduced on the SDDEnhancements
branch.