[#74774412] Get DataSource multi-rows showing up
@evanweaver the problem wasn't with the usage of the DataSource,
the problem was that moving from QWidget for the container to QCheckBox
somehow mysteriously caused the child widgets to disappear.

By moving to a shared QLayout instead of a layout containing a widget
that contains a layout that contains the child rows, the child rows now show
up.

This design really makes more sense anyhow. What doesn't make sense is that
they are all squished into one tiny space now. But I'll leave the
layout / usability fixes to you.