[#74774412] Stub in support for SpaceInfiltrationDesignFlowRate
@evanweaver I believe this does what you want, but I don't see a
method in SpaceType to tie the DropZone's setter method to for
adding a SpaceInfiltrationDesignFlowRate object, so right now
there is nothing tied to the getter and setter for the DropZone.

I would think you'd never want anything tied to the setter, you want
it to always appear empty, and if something is dropped you want it
added to the displayed list.