Brute force workaround to crash on thermalzone grid
When editing a text field on the main tab of the ThermalZone grid view,
a crash occurs. This is a brute force workaround which simply ignores
the error and continues on. It appears to work, but I don't understand
the code base enough to know if this is a proper solution

@evanweaver