Improve IG validation
Some autosizable fields were reported to not accept 0 as a value.
Change from autosize to hardsize with 0, and the IG would appear to
accept that.  But leave IG and come back and you will find that the
field is still set to autosize.

The root of this problem seems to be that some idd fields use exclusive
limits.  ie > not >=.  However the IG input field validators don't
discern the difference and therefore just default to inclusive limits.
This lets the IG interface lead you to believe you set a value that
really didn't validate and get written to the model.

This is a change to to add / subtract epsilon if the limit is exclusive,
so that you will know right away if a value is not accepted.

[fix #85606174]
fix #1286