Fix crash on remove coil from plant
When a coil is removed from a plant, a ControllerWaterCoil is also
removed by the model because it is no longer valid.  It happens that if
initiate this operation via the
inspector, the inspector's other tab (your are probably on the loop
chooser tab) is holding that controller which is
removed.  Recent improvements to inspector
(5f320d375b5925dc0f1061fb83c3a929fee48904) to make it refresh the entire
inspector (all tabs) on object removed, cause the loop chooser widget to
be removed while it is in use.  Again as a result of the controller
going away.  This fix is simply to disconnect the signal fired when the
controller is removed so that the entire IG is not reset.  The
controller IG sub tab still handles the controller remove gracefully.

This may hold the record for time spent to identify a one line bug fix.

fix #1380
[#87764568]