Fix IG bug related to auto/hard size
When you have an autosizable field, IG would give focus to the hard size
input when you clicked on the field label.  This had the undesired
consequence that the hard sized value would be used in the simulation
even if the autosize radio button was still clicked.  This is bad and
the behavior is just weird.  The solution is as easy as removing the
correct two lines of code.

close #1476