Patch additional phantom-disconnect bugs
Application now appears to run without signal-slot related crashing.

It has been noted that add/remove buttons are non-functional, and the
edit dialog on the right column bar are not present. Unsure as to
whether that is a result of the commit we inherited or if there is a
problem on our side.