Properly export symbols for built DLL's
This gets MSVC build *mostly* working. The problem now
is linking of openstudio_modeleditor_rb. There is
currently code in place in

ruby/CMakeLists.txt

to ignore missing symbols for ruby dependencies on clang
and gcc. We can attempt the same thing with -FORCE:UNRESOLVED
for the msvc linker. However, CMake still reads the errors reported
as a failure and CMake will not continue.

This can likely be worked around. However, I want to know why
we have to ignore these linker errors instead of properly linking
to a provided .lib file form the OpenStudio library?